+++ src/main/java/com/htdinh/bookstore/dto/request/ProductCartRequest.java
@@ -3,13 +3,10 @@ package com.htdinh.bookstore.dto.request;
 import lombok.Getter;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Getter
 public class ProductCartRequest {
-
-
-    BigDecimal productId;
+    Long productId;
     @NotNull
     Long quantity;
 }
