+++ src/main/java/com/htdinh/bookstore/service/ProductService.java
@@ -4,20 +4,18 @@ import com.htdinh.bookstore.dto.request.ProductRequest;
 import com.htdinh.bookstore.dto.response.ProductResponse;
 import org.springframework.data.domain.Page;
 
-import java.math.BigDecimal;
-
 public interface ProductService {
 
     Page<ProductResponse> getAllProduct(int pageNumber, int pageSize, Integer seed);
 
     Page<ProductResponse> getProductsByName(String searchKey, int pageNumber, int pageSize);
 
-    void publishProductByShop(BigDecimal productId);
+    void publishProductByShop(Long productId);
 
     Page<ProductResponse> getAllPublishForShop(int pageNumber, int pageSize);
 
 
-    void draftProductByShop(BigDecimal productId);
+    void draftProductByShop(Long productId);
 
     Page<ProductResponse> getAllDraftForShop(int pageNumber, int pageSize);
